requirements.txt
alembic==1.17.1
<<<<<<< HEAD
blinker==1.9.0
cffi==2.0.0
click==8.1.8
colorama==0.4.6
cryptography==46.0.5
ecdsa==0.19.1
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
git-filter-repo==2.47.0
greenlet==3.2.4
=======
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.129.0
git-filter-repo==2.47.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
>>>>>>> 2d6fbf2f26c40f214140ad6009db07046db5635d
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.5.0
<<<<<<< HEAD
pyasn1==0.6.2
pycparser==3.0
PyJWT==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
=======
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.12
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.44
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
>>>>>>> 2d6fbf2f26c40f214140ad6009db07046db5635d
Werkzeug==3.1.3
